Do...while
Tried using do...while for the user input section but ended up
triggering a crash, at least I think it was. The console.log showed
a number that kept increasing, frighteningly, so I just closed the
browser.

Still trying to get a prompt to show until a suitable answer is
received. So far I'm only able to console.log if the first or second
input is correct or incorrect but then it just stops there.