Filling out playGame() function
Currently the playGame() function will only console.log 'bad input'
regardless of what player enters. All possible scenarios are written
out as else...if statements.

The computerPlay() function was reworked to include computerSelection
variable. This variable is set as a parameter for playGame() but I'm
not sure if it may be contributing to the issue of the 'bad input'.