Fixed the 'bad input' issue
Had to change the ORs to ANDs in the lower part of the playGame
function. The ORs were always returning at least one 'true' which
kept triggering the console.log. Switching to ANDs makes it so that
having at least one false will trigger the else statement. This is
able to catch actual bad inputs and show the console.log.

The next part to solve is the else that gets triggered by the else if
just above it on line 38.