Function game()
Next step is to create game() function which will execute playRound()
five times and record the scores. I broke up the winner console.log
into two separate phrases because I want to see if there is a way to
count how many times 'CPU wins.' and score based on that.