Function game() continued...
The game() function executes playRound() five times successfully. It
catches bad inputs and does not count it as one of the five rounds.

The playRound() function was changed to include a function after
every outcome - this was to try to use console.count to log how many
times a certain function (win announcement) is called. I tried using
this as a way to count score but it is not working properly.