Score logging
The playRound() function now has console.log to show the score after
every round. Variables for playerScore and cpuScore were created to
track the results of these rounds.

In the game() function, after 5 rounds, the winner will be announced
or a message will show that it is a tie.

The HTML also has two lines of text explaining what will happen.