Operate function prompts user for numbers
Currently the operate() only asks the user for numbers. The numbers are
properties that are assigned to the object keys ("first" and "second").
The prompt then should return the object that the numbers are assigned
to but it returns "undefined".

Upon checking the type of data of the properties, it shows that they are
strings.