Stock picker method
This method takes in an array of hypothetical prices and chooses the
best day to buy and the best day to sell.

Currently the method can evaluate each day's best potential earnings
(Negative earnings come out to 0.) and output the string "Day X - X's
best day earns X."

Method still need to resolve this into an array of just two numbers to
complete the assignment.